+++ src/test/groovy/pl/futurecollars/invoicing/controller/AbstractControllerTest.groovy
@@ -64,12 +64,12 @@ class AbstractControllerTest extends Specification {
         }
     }
 
-    void deleteInvoice(int id) {
+    void deleteInvoice(long id) {
         mockMvc.perform(delete("$INVOICE_ENDPOINT/$id"))
                 .andExpect(status().isNoContent())
     }
 
-    Invoice getInvoiceById(int id) {
+    Invoice getInvoiceById(long id) {
         def invoiceAsString = mockMvc.perform(get("$INVOICE_ENDPOINT/$id"))
                 .andExpect(status().isOk())
                 .andReturn()
@@ -79,7 +79,7 @@ class AbstractControllerTest extends Specification {
         jsonService.toObject(invoiceAsString, Invoice)
     }
 
-    String invoiceAsJson(int id) {
+    String invoiceAsJson(long id) {
         jsonService.toJson(invoice(id))
     }
